fix: constants refactor and change .eslint config
