+++ src/common/constants/response-errors.constants.ts
@@ -1,0 +1,5 @@
+export const enum RESPONSE_ERROR_MESSAGES {
+  WRONG_POST_NOTE = 'Wrong note body',
+  WRONG_NAME_ERROR = 'Wrong input name',
+  WRONG_ID = 'Wrong id value',
+}
