+++ src/common/constants/url.constants.ts
@@ -1,0 +1,7 @@
+export const URL_PREFIX = 'api';
+
+export enum BASE_URLS {
+  GREETINGS = '/greetings/:name?',
+  NOTES = '/notes',
+  NOTES_BY_ID = '/notes/:id?',
+}
